source/requirements.txt
pydantic
sqlmodel
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
pydantic-settings
starlette